refactor(session): session_management 拆五平铺子模块（4008→317）
